[Conversation] Fix unread status when adding participant via mention approval
Move validateUserMention to mentions.ts and set unread=true when
approving a user mention, so the conversation appears in the
mentioned user's inbox.
