+++ front/lib/api/assistant/conversation/validate_actions.ts
@@ -28,11 +28,10 @@ import {
   MessageModel,
   UserMessageModel,
 } from "@app/lib/models/agent/conversation";
-import { triggerConversationAddedAsParticipantNotification } from "@app/lib/notifications/workflows/conversation-added-as-participant";
 import { AgentMCPActionResource } from "@app/lib/resources/agent_mcp_action_resource";
 import { AgentStepContentResource } from "@app/lib/resources/agent_step_content_resource";
-import { ConversationResource } from "@app/lib/resources/conversation_resource";
-import logger, { auditLog } from "@app/logger/logger";
+import type { ConversationResource } from "@app/lib/resources/conversation_resource";
+import logger from "@app/logger/logger";
 import { launchAgentLoopWorkflow } from "@app/temporal/agent_loop/client";
 import type {
   AgentMessageType,
@@ -46,7 +45,6 @@ import {
   Err,
   isAgentMessageType,
   isContentFragmentType,
-  isRichUserMention,
   isUserMessageType,
   Ok,
   toMentionType,
@@ -485,207 +483,6 @@ export async function dismissMention(
       conversation,
       updatedMessages.agentMessages
     );
-  }
-
-  return new Ok(undefined);
-}
-
-export async function validateUserMention(
-  auth: Authenticator,
-  {
-    conversationId,
-    userId,
-    messageId,
-    approvalState,
-  }: {
-    conversationId: string;
-    userId: string;
-    messageId: string;
-    approvalState: "approved" | "rejected";
-  }
-): Promise<Result<void, APIErrorWithStatusCode>> {
-  const conversationRes = await getConversation(auth, conversationId);
-  if (conversationRes.isErr()) {
-    return new Err({
-      status_code: 404,
-      api_error: {
-        type: "conversation_not_found",
-        message: "Conversation not found",
-      },
-    });
-  }
-
-  const conversation = conversationRes.value;
-
-  // Verify the message exists
-  const message = conversation.content.flat().find((m) => m.sId === messageId);
-
-  if (!message) {
-    return new Err({
-      status_code: 404,
-      api_error: {
-        type: "message_not_found",
-        message: "Message not found",
-      },
-    });
-  }
-  if (approvalState === "approved") {
-    auditLog(
-      {
-        author: auth.getNonNullableUser().toJSON(),
-        workspaceId: conversation.owner.sId,
-        conversationId: conversation.sId,
-        messageId: message.sId,
-        userId,
-        approvalState,
-      },
-      "User approved a mention"
-    );
-  }
-
-  if (isUserMessageType(message)) {
-    // Verify user is authorized to edit the message by checking the message user.
-    if (message.user && message.user.id !== auth.getNonNullableUser().id) {
-      return new Err({
-        status_code: 403,
-        api_error: {
-          type: "invalid_request_error",
-          message: "User is not authorized to edit this mention",
-        },
-      });
-    }
-  } else if (isAgentMessageType(message)) {
-    // Verify user is authorized to edit the message by going back to the user message.
-    const { userMessageUserId } = await getUserMessageIdFromMessageId(auth, {
-      messageId,
-    });
-    if (
-      userMessageUserId !== null &&
-      userMessageUserId !== auth.getNonNullableUser().id
-    ) {
-      return new Err({
-        status_code: 403,
-        api_error: {
-          type: "invalid_request_error",
-          message: "User is not authorized to edit this mention",
-        },
-      });
-    }
-  } else if (isContentFragmentType(message)) {
-    return new Err({
-      status_code: 400,
-      api_error: {
-        type: "invalid_request_error",
-        message: "Invalid message type",
-      },
-    });
-  } else {
-    assertNever(message);
-  }
-  const user = await getUserForWorkspace(auth, {
-    userId,
-  });
-  if (!user) {
-    return new Err({
-      status_code: 404,
-      api_error: {
-        type: "user_not_found",
-        message: "User not found",
-      },
-    });
-  }
-
-  const updatedMessages: {
-    userMessages: UserMessageType[];
-    agentMessages: AgentMessageType[];
-  } = {
-    userMessages: [],
-    agentMessages: [],
-  };
-  // Find all pending mentions for the same user on conversation messages latest versions.
-  for (const messageVersions of conversation.content) {
-    const latestMessage = messageVersions[messageVersions.length - 1];
-
-    if (
-      latestMessage.visibility !== "deleted" &&
-      !isContentFragmentType(latestMessage) &&
-      latestMessage.richMentions.some(
-        (m) => m.status === "pending" && m.id === userId
-      )
-    ) {
-      const mentionModel = await MentionModel.findOne({
-        where: {
-          workspaceId: conversation.owner.id,
-          messageId: latestMessage.id,
-          userId: user.id,
-        },
-      });
-      if (!mentionModel) {
-        continue;
-      }
-      await mentionModel.update({ status: approvalState });
-      const newRichMentions = latestMessage.richMentions.map((m) =>
-        isRichUserMention(m) && m.id === userId
-          ? {
-              ...m,
-              status: approvalState,
-            }
-          : m
-      );
-      if (isUserMessageType(latestMessage)) {
-        updatedMessages.userMessages.push({
-          ...latestMessage,
-          richMentions: newRichMentions,
-          mentions: newRichMentions.map(toMentionType),
-        });
-      } else if (isAgentMessageType(latestMessage)) {
-        updatedMessages.agentMessages.push({
-          ...latestMessage,
-          richMentions: newRichMentions,
-        });
-      }
-    }
-  }
-
-  for (const userMessage of updatedMessages.userMessages) {
-    await publishMessageEventsOnMessagePostOrEdit(
-      conversation,
-      {
-        ...userMessage,
-        contentFragments: getRelatedContentFragments(conversation, userMessage),
-      },
-      []
-    );
-  }
-
-  if (updatedMessages.agentMessages.length > 0) {
-    await publishAgentMessagesEvents(
-      conversation,
-      updatedMessages.agentMessages
-    );
-  }
-
-  const isParticipant = await ConversationResource.isConversationParticipant(
-    auth,
-    {
-      conversation,
-      user: user.toJSON(),
-    }
-  );
-
-  if (!isParticipant && approvalState === "approved") {
-    const status = await ConversationResource.upsertParticipation(auth, {
-      conversation,
-      action: "subscribed",
-      user: user.toJSON(),
-    });
-
-    if (status === "added") {
-      await triggerConversationAddedAsParticipantNotification(auth, {
-        conversation,
-        addedUserId: user.sId,
-      });
-    }
   }
 
   return new Ok(undefined);
